caches: Introduce cache_hit and cache_refresh helpers
Along the way a struct cacheline_desc is used to describe
a cacheline (tag, idx and cpu).

This change paves the way for upcoming changes, ie cache
associativity. Note that the check for a hit/miss is
now abstracted with a function, which will allow us
to play with associativity there in a clean way.

Signed-off-by: Emilio G. Cota <cota@braap.org>