caches: support configurable associativity, with LRU eviction
This allows us to miss less, but at a greater cost per hit.
I'd need to get an estimate on the added latency for a hit;
apart from that, seems to work on !COMBINED (COMBINED seems
to be broken with FULL_CACHES since it was merged..)

Even though we hit less, the added processing penalty
means that the simulation actually runs slower (in the
host's wall time the kernel takes 4m56s with assoc=2
in full cache mode, and 4m26s with assoc=2)--but arguably
we're closer to our eventual system. Will it be worth it?
Haven't benchmarked that yet. I should also try to speed
up the hit fastpath.

Signed-off-by: Emilio G. Cota <cota@braap.org>