caches: remove idx bits from the tag
In hardware, tags and indexes would never share any bits,
since that would be a waste of area. The appended converts
our software implementation to do just that.

This shouldn't impact performance, but it's good to have it
to make the code easy to understand for someone with hardware
experience implementing caches.

Signed-off-by: Emilio G. Cota <cota@braap.org>