target-arm: support cycle counter for arm11mp (dirty hack inside!)
Note that here we bend the ISA a little; instead of providing interrupts
(for these to work we'd need to add an external PMU controller, and
fix the affinity tuning in QEMU and our kernel code) we just provide
low/high registers for perf to access them.

For all our tests 64-bit regs shouldn't overflow.

perf has been modified to accomodate the 64-bit hack.

Signed-off-by: Emilio G. Cota <cota@braap.org>