target-arm: add memory exclusive (memex) helpers
The current memory_exclusive handling is rather insane; it only
works reliably for FULL_CACHES, and it does so in a very ugly
way: systemc is invoked and it eventually calls a function in the
the main program that instantiated systemc itself (phew!)

The implementation for late or no caches is admittedly broken,
and this has become apparent with the recent exclusive ops
fixes.

The appended implements the tracking of lines acquired by the
exclusive ops, by keeping the requests in a linked list.
This is not terribly efficient, but seems correct, which is
what I'm concerned with right now. We'll have time (or not) to
make it faster.

The list implementation is taken from the kernel.

Signed-off-by: Emilio G. Cota <cota@braap.org>